refactor: Remove index in components, Change import statement
