--- src/components/Headers/index.js
@@ -1,3 +1,0 @@
-import GameHeader from "./GameHeader";
-
-export { GameHeader };
