+++ src/components/Layouts/GameLayout.js
@@ -5,7 +5,7 @@ import { View, StyleSheet } from "react-native";
 import { SafeAreaView } from "react-native-safe-area-context";
 import PropTypes from "prop-types";
 
-import { GameHeader } from "..";
+import GameHeader from "../Headers/GameHeader";
 import { navigations, colors } from "../../common/constants";
 
 const GameLayout = ({
