--- src/components/Layouts/index.js
@@ -1,4 +1,0 @@
-import Carousel from "./Carousel";
-import GameLayout from "./GameLayout";
-
-export { GameLayout, Carousel };
