+++ src/components/Modals/ContentModal.js
@@ -3,7 +3,7 @@ import React from "react";
 import { Modal, StyleSheet, Text, View } from "react-native";
 import PropTypes from "prop-types";
 
-import { Carousel } from "..";
+import Carousel from "../Layouts/Carousel";
 import { colors } from "../../common/constants";
 
 const ContentModal = ({
