+++ src/components/Modals/FailModalContainer.js
@@ -3,7 +3,7 @@ import React, { useState } from "react";
 import { useNavigation } from "@react-navigation/native";
 import PropTypes from "prop-types";
 
-import { ButtonModal } from "..";
+import ButtonModal from "./ButtonModal";
 import SmallButton from "../Buttons/SmallButton";
 import { navigations, colors } from "../../common/constants";
