+++ src/components/Modals/NextStageModalContainer.js
@@ -2,7 +2,7 @@ import React, { useState } from "react";
 
 import PropTypes from "prop-types";
 
-import { ButtonModal } from "..";
+import ButtonModal from "./ButtonModal";
 import SmallButton from "../Buttons/SmallButton";
 import { colors } from "../../common/constants";
