--- src/components/Modals/index.js
@@ -1,17 +1,0 @@
-import ButtonModal from "./ButtonModal";
-import ContentModal from "./ContentModal";
-import FailModalContainer from "./FailModalContainer";
-import InputModal from "./InputModal";
-import InputModalContainer from "./InputModalContainer";
-import NextStageModalContainer from "./NextStageModalContainer";
-import ResetModalContainer from "./ResetModalContainer";
-
-export {
-  ButtonModal,
-  ContentModal,
-  FailModalContainer,
-  InputModal,
-  InputModalContainer,
-  NextStageModalContainer,
-  ResetModalContainer,
-};
