--- src/components/Timers/index.js
@@ -1,3 +1,0 @@
-import TextTimer from "./TextTimer";
-
-export { TextTimer };
