--- src/components/index.js
@@ -1,30 +1,0 @@
-import ChanceIcons from "./ChanceIcons";
-import { GameHeader } from "./Headers";
-import { GameLayout, Carousel } from "./Layouts";
-import Logo from "./Logo";
-import {
-  ButtonModal,
-  ContentModal,
-  FailModalContainer,
-  InputModal,
-  InputModalContainer,
-  NextStageModalContainer,
-  ResetModalContainer,
-} from "./Modals";
-import { TextTimer } from "./Timers";
-
-export {
-  Logo,
-  ButtonModal,
-  ContentModal,
-  FailModalContainer,
-  InputModal,
-  InputModalContainer,
-  NextStageModalContainer,
-  ResetModalContainer,
-  GameHeader,
-  GameLayout,
-  Carousel,
-  TextTimer,
-  ChanceIcons,
-};
